Creates dedicated testing page
